+++ src/main/java/com/library/config/DatabaseConfig.java
@@ -1,6 +1,6 @@
 package com.library.config;
 
-import com.library.userRequests.userRequests;
+import com.library.userRequests.UserRequests;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
@@ -10,9 +10,6 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.HashMap;
 
-/**
- * Created by Attila on 16/08/2016.
- */
 @Configuration
 public class DatabaseConfig {
 
@@ -42,7 +39,7 @@ public class DatabaseConfig {
     }
 
     @Bean
-    public userRequests userRequest() {
-        return new userRequests(dataSource());
+    public UserRequests userRequest() {
+        return new UserRequests(dataSource());
     }
 }
