+++ src/main/java/com/library/controller/LoginController.java
@@ -1,14 +1,10 @@
 package com.library.controller;
 
-import com.library.model.User;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
-/**
- * Created by Attila on 16/08/2016.
- */
 @Controller
 public class LoginController {
