+++ src/main/resources/static/css/style.css
@@ -1,7 +1,9 @@
- body {
-    background: black;
- }
-
  .container{
      width: 40%;
      }
+
+ body {
+    background-image: url("background.jpg");
+    background-repeat: no-repeat;
+    background-size: cover;
+ }
