+++ src/main/resources/templates/registration.html
@@ -4,14 +4,8 @@
     <title>User registration</title>
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"/>
+    <link th:href="@{/css/style.css}" rel="stylesheet" href="../css/style.css"/>
 </head>
-
-<style>
-    .container{
-    width: 40%;
-    }
-
-</style>
 
 <body>
